backend/requirements.txt
fastapi
uvicorn
tinydb
pydantic
PyJWT==2.8